% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runApp.R
\name{infercnvApp}
\alias{infercnvApp}
\title{Run infercnv shiny app}
\usage{
infercnvApp()
}
\value{
The Shiny App Runs
}
\description{
This function is used to run the infercnv app.
}
\examples{
if(interactive()){
  infercnvApp()
}

}
